+++ Microsoft.Azure.Cosmos/src/Util/Extensions.cs
@@ -38,6 +38,7 @@ namespace Microsoft.Azure.Cosmos
 
             CosmosClientSideRequestStatistics cosmosClientSideRequestStatistics = response.RequestStats as CosmosClientSideRequestStatistics;
             cosmosResponse.Diagnostics = new PointOperationStatistics(
+                activityId: cosmosResponse.Headers.ActivityId,
                 statusCode: response.StatusCode,
                 subStatusCode: response.SubStatusCode,
                 requestCharge: cosmosResponse.Headers.RequestCharge,